tryout backport
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@49613 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/linux/universal/linux-3.18/net/ipv6/inet6_hashtables.c
@@ -307,7 +307,7 @@ not_unique:
 	return -EADDRNOTAVAIL;
 }
 
-static inline u32 inet6_sk_port_offset(const struct sock *sk)
+static inline u64 inet6_sk_port_offset(const struct sock *sk)
 {
 	const struct inet_sock *inet = inet_sk(sk);
